Add logcat operation through IPC

+++ madbfs-common/src/ipc.cpp
@@ -47,6 +47,8 @@ namespace madbfs::ipc
                 return Op{ op::SetCacheSize{ .mib = json::value_to<u32>(json.at("value")) } };
             } else if (op == op::names::set_ttl) {
                 return Op{ op::SetTTL{ .sec = json::value_to<i32>(json.at("value")) } };
+            } else if (op == op::names::logcat) {
+                return op::Logcat{ .color = json::value_to<bool>(json.at("value")) };
             }
 
             return Unexpect{ fmt::format("'{}' is not a valid operation, try 'help'", op) };
@@ -55,6 +57,33 @@ namespace madbfs::ipc
         } catch (...) {
             return Unexpect{ "unknown error" };
         }
+    }
+}
+
+namespace madbfs::ipc
+{
+    std::deque<LogcatSink::Msg>& LogcatSink::swap()
+    {
+        auto new_index = (m_index + 1) % 2;
+        return m_queue[std::exchange(m_index, new_index)];
+    }
+
+    void LogcatSink::sink_it_(const spdlog::details::log_msg& msg)
+    {
+        auto buf = spdlog::memory_buf_t{};
+
+        msg.color_range_start = 0;
+        msg.color_range_end   = 0;
+
+        this->formatter_->format(msg, buf);
+
+        auto& queue = m_queue[m_index];
+        if (queue.size() == m_max_queue) {
+            queue.pop_front();
+        }
+        queue.emplace_back(
+            fmt::to_string(buf), msg.color_range_start, msg.color_range_end, static_cast<usize>(msg.level)
+        );
     }
 }
 
@@ -79,24 +108,27 @@ namespace madbfs::ipc
         return Client{ path.c_str(), std::move(sock) };
     }
 
-    AExpect<json::value> Client::send(Op op)
+    void Client::stop()
+    {
+        m_socket.cancel();
+        m_socket.close();
+    }
+
+    AExpect<json::value> Client::send(FsOp op)
     {
         namespace n = op::names;
 
-        auto overload = util::Overload{
-            // clang-format off
-            [&](op::Help           ) { return json::value{ { "op", n::help             }                      }; },
+        // clang-format off
+        auto op_json = op.visit(util::Overload{
             [&](op::Info           ) { return json::value{ { "op", n::info             }                      }; },
             [&](op::InvalidateCache) { return json::value{ { "op", n::invalidate_cache }                      }; },
             [&](op::SetPageSize  op) { return json::value{ { "op", n::set_page_size    }, { "value", op.kib } }; },
             [&](op::SetCacheSize op) { return json::value{ { "op", n::set_cache_size   }, { "value", op.mib } }; },
             [&](op::SetTTL       op) { return json::value{ { "op", n::set_ttl          }, { "value", op.sec } }; },
-            // clang-format on
-        };
+        });
+        // clang-format on
 
-        auto op_json = std::visit(std::move(overload), op);
-        auto res     = co_await send_message(m_socket, json::serialize(op_json));
-        if (not res) {
+        if (auto res = co_await send_message(m_socket, json::serialize(op_json)); not res) {
             co_return Unexpect{ res.error() };
         }
 
@@ -112,6 +144,40 @@ namespace madbfs::ipc
             co_return Unexpect{ Errc::bad_message };
         }
     }
+
+    AExpect<json::value> Client::help()
+    {
+        auto op_json = json::value{ { "op", op::names::help } };
+        if (auto res = co_await send_message(m_socket, json::serialize(op_json)); not res) {
+            co_return Unexpect{ res.error() };
+        }
+
+        auto response_str = co_await receive_message(m_socket);
+        if (not response_str) {
+            co_return Unexpect{ response_str.error() };
+        }
+
+        try {
+            co_return json::parse(*response_str);
+        } catch (const std::exception& e) {
+            log_e("{}: failed to deserialize response: {}", __func__, e.what());
+            co_return Unexpect{ Errc::bad_message };
+        }
+    }
+
+    AExpect<Gen<AExpect<String>>> Client::logcat(op::Logcat opt)
+    {
+        auto op_json = json::value{ { "op", op::names::logcat }, { "value", opt.color } };
+        if (auto res = co_await send_message(m_socket, json::serialize(op_json)); not res) {
+            co_return Unexpect{ res.error() };
+        }
+
+        co_return [sock = &m_socket](this auto) -> Gen<AExpect<String>> {
+            while (sock->is_open()) {
+                co_yield receive_message(*sock);
+            }
+        }();
+    }
 }
 
 namespace madbfs::ipc
@@ -121,8 +187,10 @@ namespace madbfs::ipc
         if (m_running) {
             stop();
         }
-        if (auto sock = m_socket_path.c_str(); ::unlink(sock) < 0) {
-            log_e("{}: failed to unlink socket: {} [{}]", __func__, sock, strerror(errno));
+        if (not m_socket_path.empty()) {
+            if (auto sock = m_socket_path.c_str(); ::unlink(sock) < 0) {
+                log_e("{}: failed to unlink socket: {} [{}]", __func__, sock, strerror(errno));
+            }
         }
     }
 
@@ -147,12 +215,28 @@ namespace madbfs::ipc
         return Server{ path.c_str(), std::move(acc) };
     }
 
-    Await<void> Server::launch(OnOp on_op)
+    Await<void> Server::launch(OnFsOp on_op)
     {
         log_d("{}: ipc launched!", __func__);
+
         m_running = true;
         m_on_op   = std::move(on_op);
-        co_await run();
+
+        if (not m_logcat_sink) {
+            m_logcat_sink = std::make_shared<LogcatSink>();
+            m_logcat_sink->set_level(log::Level::off);
+            m_logcat_sink->set_pattern(log::logger_pattern);
+
+            if (auto logger = spdlog::get(log::logger_name); not logger) {
+                log_e("{}: can't find logger with name '{}'", __func__, log::logger_name);
+                co_return;
+            } else {
+                auto& sinks = logger->sinks();
+                sinks.push_back(m_logcat_sink);
+            }
+        }
+
+        std::ignore = co_await async::wait_all(run(), logcat_handler());
     }
 
     void Server::stop()
@@ -160,6 +244,7 @@ namespace madbfs::ipc
         m_running = false;
         m_socket.cancel();
         m_socket.close();
+        m_logcat_timer.cancel();
     }
 
     Await<void> Server::run()
@@ -178,6 +263,20 @@ namespace madbfs::ipc
 
     Await<void> Server::handle_peer(Socket sock)
     {
+        // clang-format off
+        static const auto help_json = json::value{
+            { "operations", {
+                ipc::op::names::help,
+                ipc::op::names::logcat,
+                ipc::op::names::info,
+                ipc::op::names::invalidate_cache,
+                ipc::op::names::set_page_size,
+                ipc::op::names::set_cache_size,
+                ipc::op::names::set_ttl,
+            } },
+        };
+        // clang-format on
+
         auto op_str = co_await receive_message(sock);
         if (not op_str) {
             co_return;
@@ -191,7 +290,16 @@ namespace madbfs::ipc
 
         if (op) {
             status = "success";
-            value  = co_await m_on_op(*op);
+
+            switch (op->index()) {
+            case 0: /* fs_op  */ value = co_await m_on_op(*op->as<FsOp>()); break;
+            case 1: /* help   */ value = help_json; break;
+            case 2: /* logcat */ {
+                m_logcat_subscribers.emplace_back(std::move(sock), op->as<op::Logcat>()->color);
+                co_return;
+            }
+            default: log_c("{}: [BUG] not all op variants are handled!", __func__); co_return;
+            }
         } else {
             status = "error";
             value  = op.error();
@@ -206,5 +314,90 @@ namespace madbfs::ipc
             const auto msg = std::make_error_code(res.error()).message();
             log_w("{}: failed to send message: {}", __func__, msg);
         }
+    }
+
+    Await<void> Server::logcat_handler()
+    {
+        using namespace std::chrono_literals;
+
+        log_i("{}: start", __func__);
+
+        // copying what spdlog do on ansicolor_sink :P
+        const auto colors = Array<Str, log::Level::n_levels>{
+            "\033[37m",           // trace    | white
+            "\033[36m",           // debug    | cyan
+            "\033[32m",           // info     | green
+            "\033[33m\033[1m",    // warn     | yellow_bold
+            "\033[31m\033[1m",    // error    | red bold
+            "\033[1m\033[41m",    // critical | bold on red
+            "\033[m",             // off      | reset
+        };
+
+        auto inactive_subscribers = Vec<isize>{};
+        auto prev_empty           = m_logcat_subscribers.empty();
+
+        while (m_running) {
+            m_logcat_timer.expires_after(100ms);
+            co_await m_logcat_timer.async_wait();
+
+            if (m_logcat_subscribers.empty()) {
+                if (not prev_empty) {
+                    log_i("{}: no subscribers, turning off logcat sink", __func__);
+                    m_logcat_sink->set_level(log::Level::off);
+                }
+                prev_empty = true;
+                continue;
+            } else if (prev_empty) {
+                log_i("{}: subscribers not empty anymore, turning on logcat sink", __func__);
+                m_logcat_sink->set_level(spdlog::get(log::logger_name)->level());
+                prev_empty = false;
+            }
+
+            auto& messages = m_logcat_sink->swap();
+            auto  buffer   = String{};
+
+            for (auto&& [i, sub] : m_logcat_subscribers | sv::enumerate) {
+                auto& [sock, should_color] = sub;
+                for (auto& [message, col_start, col_end, level] : messages) {
+                    auto res = Expect<void>{};
+
+                    if (should_color) {
+                        auto it = message.begin();
+                        buffer.append(it, it + col_start);              // before
+                        buffer.append(colors.at(level));                // color code
+                        buffer.append(it + col_start, it + col_end);    // color range
+                        buffer.append(colors.back());                   // reset code
+                        buffer.append(it + col_end, message.end());     // after
+                        res = co_await send_message(sock, buffer);
+                        buffer.clear();
+                    } else {
+                        res = co_await send_message(sock, message);
+                    }
+
+                    if (not res) {
+                        auto err_msg = std::make_error_code(res.error()).message();
+                        log_w("{}: peer error: {}", __func__, i);
+                        inactive_subscribers.emplace_back(i);
+                        break;
+                    }
+                }
+            }
+
+            for (auto idx : inactive_subscribers | sv::reverse) {
+                m_logcat_subscribers.erase(m_logcat_subscribers.begin() + idx);
+            }
+
+            inactive_subscribers.clear();
+            messages.clear();
+        }
+
+        for (auto& [sock, _] : m_logcat_subscribers) {
+            sock.cancel();
+            sock.close();
+        }
+
+        m_logcat_subscribers.clear();
+
+        log_i("{}: end", __func__);
     }
 }
